requirements2.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
azure-common==1.1.28
azure-core==1.30.2
azure-search==1.0.0b2
azure-search-documents==11.5.1
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
idna==3.8
isodate==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-experimental==0.3.3
langchain-openai==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.1.140
marshmallow==3.22.0
msrest==0.7.1
multidict==6.0.5
mypy-extensions==1.0.0
oauthlib==3.2.2
openai==1.43.0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pydantic==2.8.2
pydantic-settings==2.6.1
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlparse==0.5.1
tenacity==8.5.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
yarl==1.9.7